test(og): keep Sharp rasterize out of Cypress CT
The CT shard imported installOgFonts and sharp in the browser bundle.
Webpack then crashed on fs.existsSync and the TTF, so the whole spec
failed before any assertion ran.

Keep the SVG string checks in Cypress. Move the font/pixel check to a
Node mocha spec and inject the Lato data URI at rasterize time.
