+++ ui/cypress/integration/unit/og-preview.cy.ts
@@ -12,9 +12,7 @@ import {
   previewOrigin,
   sanitizeDiscordEmbeds,
 } from '@/lib/og/preview'
-import { ogFontIsBundled, rasterizeOgSvg } from '@/lib/og/rasterize'
 import { escapeXml, renderOgSvg, wrapText } from '@/lib/og/svg'
-import sharp from 'sharp'
 
 describe('clip', () => {
   it('collapses whitespace and ellipsizes long strings', () => {
@@ -239,33 +237,6 @@ describe('OG SVG card', () => {
     expect(svg).to.include('moondao.com/jobs')
     expect(svg).to.include('width="1200"')
     expect(svg).to.include('height="630"')
-    expect(svg).to.include('font-family:Lato')
     expect(svg).to.include('font-family="Lato, sans-serif"')
-  })
-})
-
-describe('OG rasterize', () => {
-  it('bundles Lato and paints readable title pixels', async () => {
-    expect(ogFontIsBundled()).to.equal(true)
-    const svg = renderOgSvg({
-      eyebrow: 'MoonDAO  ·  Jobs',
-      title: 'Social Media Manager',
-      subtitle: 'Executive Branch',
-      chips: ['Marketing', 'Part-time'],
-      footer: 'moondao.com/jobs',
-    })
-    const png = await rasterizeOgSvg(svg)
-    const image = sharp(png)
-    const meta = await image.metadata()
-    expect(meta.width).to.equal(1200)
-    expect(meta.height).to.equal(630)
-    expect(meta.format).to.equal('png')
-
-    // Title baseline sits at y=200. White Lato glyphs must be present; a
-    // missing-font card is only the dark gradient (channel max ~50).
-    const titleStats = await sharp(png)
-      .extract({ left: 50, top: 160, width: 700, height: 80 })
-      .stats()
-    expect(titleStats.channels[0].max).to.be.greaterThan(200)
   })
 })
