+++ ui/lib/og/rasterize.ts
@@ -1,9 +1,19 @@
-import { installOgFonts, resolveOgFontPath } from './installOgFonts'
+import { getOgFontDataUri, installOgFonts, resolveOgFontPath } from './installOgFonts'
 import { readFileSync } from 'fs'
 import { join } from 'path'
 import sharp from 'sharp'
 
 installOgFonts()
+
+/** Librsvg needs the face in the SVG itself; fontconfig alone is not enough on Vercel. */
+export function embedOgFontFace(svg: string): string {
+  const dataUri = getOgFontDataUri()
+  if (!dataUri || svg.includes('@font-face{font-family:Lato')) return svg
+  return svg.replace(
+    '<defs>',
+    `<defs>\n    <style>@font-face{font-family:Lato;src:url('${dataUri}') format('truetype');}</style>`
+  )
+}
 
 const FALLBACK_CANDIDATES = [
   join(process.cwd(), 'public/metadata-image.png'),
@@ -27,7 +37,7 @@ function readFallbackPng(): Buffer | undefined {
 export async function rasterizeOgSvg(svg: string): Promise<Buffer> {
   installOgFonts()
   try {
-    return await sharp(Buffer.from(svg)).png().toBuffer()
+    return await sharp(Buffer.from(embedOgFontFace(svg))).png().toBuffer()
   } catch (error) {
     console.error('Failed to rasterize OG card:', error)
     const fallback = readFallbackPng()
