+++ ui/scripts/.mocharc-cypress-unit.json
@@ -19,6 +19,7 @@
     "cypress/integration/unit/listing-helpers.cy.ts",
     "cypress/integration/unit/job-posting-jsonld.cy.ts",
     "cypress/integration/unit/og-preview.cy.ts",
+    "scripts/og-rasterize.test.ts",
     "cypress/integration/unit/voting.cy.tsx",
     "cypress/integration/unit/vmooney-snapshots.cy.tsx",
     "cypress/integration/unit/audit-bug-regression.cy.tsx",
